Use 16-bit move IDs in Bill's PC for boxes other than the current one
In a similar way than with species IDs, move IDs could easily overflow the
available storage of any table, with the PC system being able to store as many
as 1,120 moves across all Pokémon in it. Therefore, boxes in "cold storage"
will store the full 16-bit move indexes for all moves. The upper bits of the
index are stored in the PP of the move, while the lower byte is stored in the
usual location; this will constrain indexes to 14 bits, since there are only 6
available bits in the PP byte. Since Pokémon are fully healed when in the PC
system, the PP for each move can be losslessly restored.